calc: add :module, :mod selectors
It's useful to draw a graph for a single command in a larger repository.
Currently it was possible to achieve this via

   shared(github.com/example/cmd/x:all, github.com/example/...)

However, it's very cumbersome, so the equivalent can be done via:

   github.com/example/cmd/x:mod

If you wish to include another module to the mix, then you would need to
add one specific package from the other module, e.g. if we wanted to
include golang.org/x/sync packages, then we can do:

   (github.com/example/cmd/x:mod + golang.org/x/sync):mod